backend/requirements.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-nested-routers>=0.93.4

# SQL Server
mssql-django>=1.4
pyodbc>=5.0.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# API Documentation
drf-spectacular>=0.27.0

# Utilities
python-dateutil>=2.8.2
Pillow>=10.0.0

# Barcode
python-barcode>=0.15.1
qrcode>=7.4.2

# PDF & Printing
reportlab>=4.0.0
weasyprint>=60.0

# Excel Export
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Environment
python-dotenv>=1.0.0

# Development
ipython>=8.0.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
hypothesis>=6.0.0
